运营数据管理模块功能

+++ ruoyi-modules/txwx-webchat/src/main/java/com/txwx/webchat/domain/vo/UserTotalVo.java
@@ -10,5 +10,5 @@ public class UserTotalVo {
     private String desc;
 
     @Schema(description = "数量")
-    private Integer value;
+    private Long value;
 }
